move redis to data-access dir
